requirements.txt
mysql.connector
sshtunnel==0.1.5
boto3
requests